==> 01_bronze_layer.ipynb <==
# 01 - Bronze Layer

Ingestão de dados brutos com metadados.


import pandas as pd
from datetime import datetime
import os


os.makedirs('data/bronze', exist_ok=True)
os.makedirs('data/silver', exist_ok=True)
os.makedirs('data/gold', exist_ok=True)


# Carregar dados
df = pd.read_csv('data.csv', sep=',', encoding='ISO-8859-1', low_memory=False)
df['data_ingestao'] = datetime.now()
df['fonte_arquivo'] = 'data.csv'

print(f"Dados carregados: {df.shape[0]:,} linhas")
df.head()


# Salvar Bronze
df.to_csv('data/bronze/dados_brutos.csv', index=False)
print("Dados salvos na camada Bronze")
